requirements.txt
asgiref==3.8.1
Django==5.0.4
djangorestframework==3.15.1
packaging==24.0
pdf2image==1.17.0
pillow==10.3.0
PyPDF2==3.0.1
pytesseract==0.3.10
sqlparse==0.5.0
